refactor(storage): rename
