+++ packages/service/storage/src/index.ts
@@ -1,5 +1,5 @@
 import './route/home.js';
-import './route/update-doc.js';
+import './route/update.js';
 import {logger} from './lib/config.js';
 
 logger.logOther('..:: Alwatr Nanoservice Starter Kit ::..');
